move redirectsConfig to redirects folder and update imports
