+++ packages/admin/cms-admin/src/config/CometConfigContext.tsx
@@ -8,7 +8,7 @@ import { type ContentLanguageConfig } from "../contentLanguage/contentLanguageCo
 import { type DamConfig } from "../dam/config/damConfig";
 import { type DependenciesConfig } from "../dependencies/dependenciesConfig";
 import { type PageTreeConfig } from "../pages/pageTreeConfig";
-import { type RedirectsConfig } from "../pages/redirectsConfig";
+import { type RedirectsConfig } from "../redirects/redirectsConfig";
 import { type SiteConfigsConfig } from "../siteConfigs/siteConfigsConfig";
 import { type WarningsConfig } from "../warnings/warningsConfig";
